nest loggers/replicators/liveness under client
