+++ client/liveness/serve.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 	"net/http"
 
-	"github.com/alehechka/kube-external-sync/replicate/common"
+	"github.com/alehechka/kube-external-sync/client/replicate/common"
 	log "github.com/sirupsen/logrus"
 )
