+++ client/replicate/common/common.go
