+++ client/sync.go
@@ -1,8 +1,8 @@
 package client
 
 import (
-	"github.com/alehechka/kube-external-sync/liveness"
-	"github.com/alehechka/kube-external-sync/replicate/common"
+	"github.com/alehechka/kube-external-sync/client/liveness"
+	"github.com/alehechka/kube-external-sync/client/replicate/common"
 	log "github.com/sirupsen/logrus"
 )
